Marketplace: extract the thank-you redirect into its own hook
Move the post-install redirect machinery — the fresh-site query, destination
URL resolution, background polling, and the plugin and theme redirect effects —
out of useProductInstall into useThankYouRedirect. Verbatim code motion: the
query, conditions, and navigation are unchanged, only relocated with their
inputs passed in.
