+++ client/my-sites/marketplace/pages/marketplace-product-install/use-product-install.tsx
@@ -1,12 +1,7 @@
-import { siteByIdQuery } from '@automattic/api-queries';
-import { WPCOM_FEATURES_ATOMIC, WPCOM_FEATURES_MANAGE_PLUGINS } from '@automattic/calypso-products';
-import page from '@automattic/calypso-router';
-import { useQuery } from '@tanstack/react-query';
+import { WPCOM_FEATURES_ATOMIC } from '@automattic/calypso-products';
 import { useTranslate } from 'i18n-calypso';
 import { useEffect, useState, useMemo, useRef } from 'react';
 import { useQueryTheme } from 'calypso/components/data/query-theme';
-import { isAtomicTransferredSite } from 'calypso/dashboard/utils/site-atomic-transfers';
-import { useInterval } from 'calypso/lib/interval';
 import { waitFor } from 'calypso/my-sites/marketplace/util';
 import { useSelector, useDispatch } from 'calypso/state';
 import { initiateAtomicTransfer } from 'calypso/state/atomic/transfers/actions';
@@ -14,11 +9,7 @@ import { transferStates } from 'calypso/state/automated-transfer/constants';
 import { getAutomatedTransferStatus } from 'calypso/state/automated-transfer/selectors';
 import { getPurchaseFlowState } from 'calypso/state/marketplace/purchase-flow/selectors';
 import { MARKETPLACE_ASYNC_PROCESS_STATUS } from 'calypso/state/marketplace/types';
-import {
-	installPlugin,
-	activatePlugin,
-	fetchSitePlugins,
-} from 'calypso/state/plugins/installed/actions';
+import { installPlugin, activatePlugin } from 'calypso/state/plugins/installed/actions';
 import {
 	getPluginOnSite,
 	getStatusForPlugin,
@@ -33,11 +24,10 @@ import getUploadedPluginId from 'calypso/state/selectors/get-uploaded-plugin-id'
 import isPluginUploadComplete from 'calypso/state/selectors/is-plugin-upload-complete';
 import isSiteAutomatedTransfer from 'calypso/state/selectors/is-site-automated-transfer';
 import siteHasFeature from 'calypso/state/selectors/site-has-feature';
-import { isJetpackSite, getSiteAdminUrl } from 'calypso/state/sites/selectors';
+import { isJetpackSite } from 'calypso/state/sites/selectors';
 import {
 	initiateThemeTransfer as initiateTransfer,
 	installAndActivateTheme,
-	requestActiveTheme,
 } from 'calypso/state/themes/actions';
 import { getTheme, isThemeActive as getThemeActive } from 'calypso/state/themes/selectors';
 import {
@@ -47,6 +37,7 @@ import {
 } from 'calypso/state/ui/selectors';
 import { useDelayedCondition } from './use-delayed-condition';
 import useMarketplaceAdditionalSteps from './use-marketplace-additional-steps';
+import { useThankYouRedirect } from './use-thank-you-redirect';
 
 // The state authorizing an install is handed off asynchronously, so allow for it arriving late.
 const INSTALL_HANDOFF_GRACE_PERIOD_MS = 2000;
@@ -256,108 +247,24 @@ export function useProductInstall( {
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [ pluginUploadComplete, installedPlugin, setCurrentStep ] );
 
-	// Fetch fresh site data (including admin_url) post-transfer
-	const { data: freshSite } = useQuery( {
-		...siteByIdQuery( siteId ?? 0 ),
-		enabled: !! siteId && ( ! atomicFlow || automatedTransferStatus === transferStates.COMPLETE ),
-		refetchInterval: ( query ) =>
-			query.state.data && isAtomicTransferredSite( query.state.data ) ? false : 2000,
-		staleTime: 0,
-		refetchOnMount: 'always',
-	} );
-
-	const freshAdminUrl = freshSite?.options?.admin_url;
-	const isAtomicTransferReady = freshSite ? isAtomicTransferredSite( freshSite ) : false;
-	const pluginsUrlFresh = freshAdminUrl
-		? `${ freshAdminUrl }plugins.php?activate=true&plugin_status=active`
-		: null;
-
-	const pluginsUrlSelector = useSelector( ( state ) =>
-		getSiteAdminUrl( state, siteId, 'plugins.php?activate=true&plugin_status=active' )
-	);
-
-	// Prefer fresh URL when available; if in atomic flow, wait for fresh URL
-	const pluginsUrlFinal = atomicFlow ? pluginsUrlFresh : pluginsUrlFresh || pluginsUrlSelector;
-
-	// For marketplace plugins (e.g. sensei-pro), the atomic transfer + plugin install
-	// is initiated during checkout, not by this component. The wporg data is unavailable,
-	// so atomicFlow is never set. Once the site is atomic, poll for installed plugins
-	// so that the existing redirect (installedPlugin && pluginActive) fires.
-	const isMarketplacePluginFlow =
-		! atomicFlow &&
-		! isPluginUploadFlow &&
-		!! pluginSlug &&
-		!! freshSite?.is_wpcom_atomic &&
-		wporgPlugin?.wporg === false;
-
-	useInterval(
-		() => dispatch( fetchSitePlugins( siteId ) ),
-		isMarketplacePluginFlow && ! pluginActive ? 3000 : null
-	);
-
-	const canManagePlugins = useSelector( ( state ) => {
-		return siteHasFeature( state, selectedSite?.ID, WPCOM_FEATURES_MANAGE_PLUGINS );
-	} );
-	// Check completition of all flows and redirect to thank you page
-	useEffect( () => {
-		if (
-			// Happens in 3 cases:
-			// - Click on "Install and activate" button for any plugin on /plugins/<site_name>
-			// - Install with the help of uploading archive of a plugins
-			// - If it's simple site which doesn't support plugins, then installing and activation happens at the same time with upgrading to Business plan
-			( installedPlugin && pluginActive ) ||
-			// Transfer to atomic using a marketplace plugin
-			( atomicFlow &&
-				transferStates.COMPLETE === automatedTransferStatus &&
-				canManagePlugins &&
-				isAtomicTransferReady ) ||
-			// Transfer to atomic uploading a zip plugin
-			( uploadedPluginSlug &&
-				isPluginUploadFlow &&
-				! isAtomic &&
-				transferStates.COMPLETE === automatedTransferStatus &&
-				canManagePlugins &&
-				isAtomicTransferReady )
-		) {
-			// Require a resolved pluginsUrlFinal before redirecting
-			if ( ! pluginsUrlFinal ) {
-				return;
-			}
-			waitFor( 1 ).then( () => {
-				window.location.href = pluginsUrlFinal as string;
-			} );
-		}
-	}, [
-		pluginActive,
-		automatedTransferStatus,
-		atomicFlow,
+	useThankYouRedirect( {
+		siteId,
+		selectedSite,
+		selectedSiteSlug,
+		currentStep,
 		isPluginUploadFlow,
-		isAtomic,
-		canManagePlugins,
+		pluginSlug,
+		themeSlug,
+		wporgPlugin,
+		wpOrgTheme,
+		isThemeActive,
 		installedPlugin,
+		pluginActive,
 		uploadedPluginSlug,
-		pluginsUrlFinal,
-		isAtomicTransferReady,
-	] ); // We need to trigger this hook also when `automatedTransferStatus` changes cause the plugin install is done on the background in that case.
-
-	// Validate theme is already active
-	useEffect( () => {
-		if ( themeSlug && wpOrgTheme && isThemeActive ) {
-			waitFor( 1 ).then( () =>
-				page.redirect(
-					`/marketplace/thank-you/${ selectedSiteSlug }?themes=${ themeSlug }&hide-progress-bar`
-				)
-			);
-		}
-	}, [ themeSlug, wpOrgTheme, isThemeActive, selectedSiteSlug ] );
-
-	// Polling for theme activation status
-	useInterval(
-		() => {
-			dispatch( requestActiveTheme( siteId ) );
-		},
-		! themeSlug || currentStep === 0 || ( themeSlug && wpOrgTheme && isThemeActive ) ? null : 3000
-	);
+		atomicFlow,
+		isAtomic,
+		automatedTransferStatus,
+	} );
 
 	const steps = useMemo( () => {
 		if ( themeSlug ) {
